backend/Lambda/requirements.txt
google-streetview==1.2.9
##scikit-image==0.18.1
flask==1.1.2
Pillow==8.1.0
matplotlib==3.3.3
numpy==1.20.1
boto3==1.17.4